+++ src/orgChart/customNodes/MentorMenteeNode.tsx
@@ -1,17 +1,20 @@
 import { memo } from "react";
 import { Handle, Position } from "react-flow-renderer";
+import { Link } from "react-router-dom";
 
 export default memo(
-  ({ data }: { data: { fullName: string; image: string } }) => {
+  ({ id, data }: { id: string; data: { fullName: string; image: string } }) => {
     return (
-      <div className={"branch-box"}>
-        <div>
-          <img className={"image"} src={data.image} alt={data.fullName} />
+      <Link to={{ pathname: `/member-account/${id}` }}>
+        <div className={"branch-box"}>
+          <div>
+            <img className={"image"} src={data.image} alt={data.fullName} />
+          </div>
+          <div>{data.fullName}</div>
+          <Handle type="target" position={Position.Top} />
+          <Handle type="source" position={Position.Bottom} />
         </div>
-        <div>{data.fullName}</div>
-        <Handle type="target" position={Position.Top} />
-        <Handle type="source" position={Position.Bottom} />
-      </div>
+      </Link>
     );
   }
 );
